refactor(cli): extract git + repo-id helpers into work_commands_git.rs (card c0bd865c phase 3 / c63a1b6e)
Phase 3 of the 5-way split. The local-git helpers that feed claim + PR-open move to a sibling module:

  Extracted to work_commands_git.rs:
    - spawn_claim_worktree — the d1b2798d + 59243bee worktree-on-claim flow
    - cwd_github_repo_id + parse_github_repo_id — origin URL → owner/repo
    - slugify — card title → git-safe branch slug
    - git_rev_parse_branch — current branch from a worktree
    - git_show_format — HEAD subject/body for the title-from-slug fix (13131f1c)

These are pure subprocess wrappers around `git`. Card dec35ec7 will migrate them to the typed `GitClient` trait from `airc-lib::tools` once that lands; until then they shell out as before.

  work_commands.rs:     1634 → 1634 LOC (-0 in this view; net -189 across phases 1+3)
  work_commands_git.rs:    0 →  205 LOC (new)

Wait — work_commands.rs still shows 1634 because phase 2 (5aa9e780/lifecycle) hasn't merged yet. Cumulative state once phases 1+2+3 land:
  Original:                          1823 LOC
  - phase 1 (#1042, print/render):   -238
  - phase 2 (#1046, lifecycle):      -374
  - phase 3 (this PR, git):          -189
  - Final after 3 phases:            ~1022 LOC

Phases 4 (gh) and 5 (review) bring it under ~700.

  - run_claim (in lifecycle / phase 2) → spawn_claim_worktree (this PR)
  - open_pr_and_link (in work_commands.rs, future phase 4) → git_show_format, git_rev_parse_branch (this PR)

All extracted fns are `pub(crate)` so cross-module callers reach them. Phase 4 (gh) will pull `open_pr_and_link` out next, after which the gh module calls `work_commands_git::git_show_format` directly without going through the umbrella.

  cargo test work_commands:: → 19 pass (including parse_github_repo_id_handles_https_ssh_and_trailing_git, slugify_lowercases_alphanumeric_collapses_separators_and_bounds_length)
  cargo clippy -p airc-cli --bin airc -- -D warnings → clean
  cargo fmt --check → clean

Phase 2 (#1046) extracts lifecycle (which calls spawn_claim_worktree). This PR (phase 3) extracts spawn_claim_worktree itself. The two are sequencing-friendly: phase 2 routes calls through `crate::work_commands::spawn_claim_worktree` (which is `pub(crate)` in work_commands.rs); when phase 3 lands, that re-export points at `crate::work_commands_git::spawn_claim_worktree`.

c0bd865c follow-up sequence. Phase 1 = render (#1042). Phase 2 = lifecycle (#1046). Phase 3 = git (this PR). Phase 4 = gh (c7bdf2df). Phase 5 = review (a3ede6de).
